requirements.txt
feedparser
pymongo
python-dotenv
beautifulsoup4
fastapi
uvicorn[standard]
httpx
cohere
langdetect
pytest
mongomock
readability-lxml
lxml
